move target JSON (de)serialization to separate file
